Deprecate AWS profiles; embrace AWS STS
I've _finally_ deprecated the use of AWS profiles. Almost all tools and
APIs that interact with AWS look for the environment variables, so
there's no reason to support them when they can be obtained from the
terminal or from a secure location, like 1Password.

This also adds support for configuring a shell with creds generated by
AWS STS. This change was LONG overdue, but it completely removes the
need for having superuser IAM users.

Lastly, I've changed my PS1 so that it always displays which AWS access
key/assumed role ARN and Azure SPN I'm using in a more convenient
format.