OMG. `ctags` and `cscope` are FUCKING AWESOME.
`ctags` can help you find the definition of a function. When used with
`g]` in vim, you basically get "Jump To Definition" without an LSP.

`cscope` is `ctags` on ACTUAL CRACK COCAINE. You can not only use it to
find the definition of a function, but you can see callers and callees
as well. You can even use `cscope` to find callers and callees of
functions that match regexps!

This commit adds a hook that allows us to build a `cscope` and `ctags`
database for any project tracked by Git. Ruby, Golang, Java, Python,
C/C++, Bash, and JavaScript are supported.

Signed-off-by: Carlos Nunez <ncarlos@vmware.com>