[feat] Test bench for Marcus
This adds a test bench for Marcus that validates its configuration
no matter where it's living.

Using Flatcar for Marcus v2 was the inspiration for this. Given that
everything in Flatcar is configured via Butane/Ignition, I figured that
this was enough of an excuse to finally make the stack that powers
everything in our house more reliable.

Signed-off-by: Carlos Nunez <13461447+carlosonunez@users.noreply.github.com>