[vim] Finally use YCM for autocomplete
I've balked at installing YouCompleteMe, a proper LSP engine, within Vim
for years, mostly due to the terrible performance overhead on Intel Macs
souring the experience.

Apple Silicon is much faster, fortunately, and it's pretty easy to
enable LSPs for languages that YCM doesn't support out of the box.

Therefore, YCM is back on the menu!

Signed-off-by: Carlos Nunez <13461447+carlosonunez@users.noreply.github.com>