[install] Remove krisp
I do not approve of how aggressively Krisp has been pushing their AI
Note taker. There's no _real_ way of disabling it (you can disable it by
turning it off in the app, but you can't remove it from the app
completely) and it nags you when you don't give it the permissions it
needs to function.

macOS and iOS have noise suppression built in ("Voice Isolation" mode)
which works with Firefox and OBS, so I'm just going to use that.

Signed-off-by: Carlos Nunez <13461447+carlosonunez@users.noreply.github.com>