[new-script] AWS: automatically set up STS user
this script automatically creates a user that can assume an admin role
in an AWS account when given a 1Password login item with the following
fields:

- `root_access_key`: An access key that has enough permissions to
  read/write into IAM
- `root_secret_key`: The secret key for that access key.

Context
========

Red Hat gives us the ability to provision blank AWS environments on
demand. However, they expire, and when that happens, the account is
deleted.

While they give us an account that has an `AdministratorAccess` policy
applied onto it, this is super-insecure (I don't know what permissions
boundaries are applied to these accounts, if any) and I'd rather use a
non-permissioned account that can assume a role, as per best practice.

To avoid doing this through the GUI and modifying 1Pass myself, I built
this to do it for me. All I need to do is update the properties above,
which the changes to `.bash_1password_specific` simplify as well.

Signed-off-by: Carlos Nunez <13461447+carlosonunez@users.noreply.github.com>